+++ src/ApiManagement/ApiManagement/ChangeLog.md
@@ -26,7 +26,7 @@
     - Added support for the new `Isolated` SKU
     - Added support for managing Availability Zones using `Zone` property
     - Added support for Disabling Gateway in a Region using `DisableGateway` property
-    - Added support for managing the minimum Api Version to allow for Control Plane using `ApiVersionConstraint` property.
+    - Added support for managing the minimum Api Version to allow for Control Plane using `MinimalControlPlaneApiVersion` property.
 
 * Updated cmdlet **New-AzApiManagementRegion** to manage ApiManagement service     
     - Added support for managing Availability Zones using `Zone` property